GH-37563: [Ruby] Unify tests about basic arrays for raw_records and each_raw_record
This PR extracted common test cases to reduce
duplications and ensure consistency between
`raw_records` and `each_raw_record`.

This PR is first step of them for removing duplicated
test cases.
